+++ fiftyone/core/odm/__init__.py
@@ -21,7 +21,6 @@ from .document import (
     SampleDocument,
     SerializableDocument,
 )
-from .fields import DictField, Field, FrameNumberField, StringField
 from .frame import (
     DatasetFrameSampleDocument,
     NoDatasetFrameSampleDocument,
